--- src/html/index.html
@@ -1,15 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-  <meta charset="utf-8">
-  <title>ToDo App</title>
-  <meta name="viewport" content="width=device-width,initial-scale=1">
-</head>
-
-<body class="container-fluid">
-  <div id="content">
-  </div>
-</body>
-
-</html>
